Refactor extending metadata-provider so it's actually working.
